the no network problem is quiet on my side
